requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.0.0
sqlparse==0.5.3
django-phonenumber-field==7.2.2
phonenumbers==8.13.45
psycopg2-binary==2.9.10
gunicorn==23.0.0
